Move LabelledDropdown local to usage
