+++ osu.Game.Tournament/Screens/SetupScreen.cs
@@ -2,6 +2,7 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using System.Collections.Generic;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -97,6 +98,26 @@ namespace osu.Game.Tournament.Screens
                     Items = rulesets.AvailableRulesets,
                     Current = LadderInfo.Ruleset,
                 },
+            };
+        }
+
+        public class LabelledDropdown<T> : LabelledComponent<OsuDropdown<T>, T>
+        {
+            public LabelledDropdown()
+                : base(true)
+            {
+            }
+
+            public IEnumerable<T> Items
+            {
+                get => Component.Items;
+                set => Component.Items = value;
+            }
+
+            protected override OsuDropdown<T> CreateComponent() => new OsuDropdown<T>
+            {
+                RelativeSizeAxes = Axes.X,
+                Width = 0.5f,
             };
         }
